fix: improve error handling and logging in system mapping and unmapping (#204)

+++ integration/helper_test.go
@@ -251,6 +251,9 @@ func deleteSystemInDB(ctx context.Context, db *gorm.DB, externalID, systemType s
 	if err != nil {
 		return err
 	}
+	if sys == nil {
+		return nil
+	}
 
 	_, err = repo.Delete(ctx, sys)
 	return err
